bd_neo4j/requirements.txt
flask
neo4j
python-dotenv
itertools
